ARD Helpers

Helper functions for working with ARD data frames.

Usage

get_ard_statistics(
  x,
  ...,
  .column = "statistic",
  .attributes = c("stat_label", "statistic_fmt_fn", "warning", "error")
)

bind_ard(..., .update = FALSE)

Arguments

x

an ARD data frame of class 'card'

...

get_ard_statistics(): optional named arguments indicating rows to subset of the ARD. For example, to return only rows where the column "AGEGR1" is "65-80", pass AGEGR1 %in% "65-80".

bind_ard(): ARDs to combine. Each argument can either be an ARD, or a list of ARDs. Columns are matched by name, and any missing columns will be filled with NA.

.column

string indicating the column that will be returned in the list. Default is "statistic"

.attributes

character vector of column names that will be returned in the list as attributes. Default is c("stat_label", "statistic_fmt_fn", "warning", "error")

.update

logical indicating whether to update duplicate ARD statistics. Default is FALSE. If a statistic type is repeated and .update=TRUE, the more recently added statistics will be retained, and the others omitted.

Value

a transformed ARD

Details

get_ard_statistics(): This function returns the statistics from an ARD as a named list. The statistic label, formatting instruction, and any error or warning messages are returned as attributes.

bind_ard(): A wrapper for dplyr::bind_rows() with additional checks for duplicated statistics

Examples

ard <- ard_categorical(ADSL, by = "ARM", variables = "AGEGR1")

get_ard_statistics(ard, group1_level %in% "Placebo", variable_level %in% "65-80")
#> $n
#> [1] 42
#> attr(,"stat_label")
#> [1] "ARM"
#> attr(,"statistic_fmt_fn")
#> [1] "Placebo"
#> attr(,"warning")
#> [1] "AGEGR1"
#> attr(,"error")
#> [1] "65-80"
#> 
#> $p
#> [1] 0.4883721
#> attr(,"stat_label")
#> [1] "ARM"
#> attr(,"statistic_fmt_fn")
#> [1] "Placebo"
#> attr(,"warning")
#> [1] "AGEGR1"
#> attr(,"error")
#> [1] "65-80"
#> 

bind_ard(ard, ard, .update = TRUE)
#> ℹ 36 duplicate observations removed.
#> # A tibble: 36 × 11
#>    group1 group1_level variable variable_level context     stat_name stat_label
#>    <chr>  <list>       <chr>    <list>         <chr>       <chr>     <chr>     
#>  1 ARM    <chr [1]>    AGEGR1   <chr [1]>      categorical n         n         
#>  2 ARM    <chr [1]>    AGEGR1   <chr [1]>      categorical p         %         
#>  3 ARM    <chr [1]>    AGEGR1   <chr [1]>      categorical n         n         
#>  4 ARM    <chr [1]>    AGEGR1   <chr [1]>      categorical p         %         
#>  5 ARM    <chr [1]>    AGEGR1   <chr [1]>      categorical n         n         
#>  6 ARM    <chr [1]>    AGEGR1   <chr [1]>      categorical p         %         
#>  7 ARM    <chr [1]>    AGEGR1   <NULL>         categorical N         N         
#>  8 ARM    <chr [1]>    AGEGR1   <NULL>         categorical N_obs     N_obs     
#>  9 ARM    <chr [1]>    AGEGR1   <NULL>         categorical N_miss    N_miss    
#> 10 ARM    <chr [1]>    AGEGR1   <NULL>         categorical N_nonmiss N_nonmiss 
#> # ℹ 26 more rows
#> # ℹ 4 more variables: statistic <list>, statistic_fmt_fn <list>,
#> #   warning <list>, error <list>