Categorical ARD Statistics

Compute Analysis Results Data (ARD) for categorical summary statistics.

Usage

ard_categorical(data, variables, by = NULL, strata = NULL, denominator = NULL)

Arguments

data

a data frame

variables

columns to include in summaries. Default is everything().

by, strata

columns to by/stratified by for tabulation. Arguments are similar, but with an important distinction:

by: results are tabulated by all combinations of the columns specified, including unobserved combinations and unobserved factor levels.

strata: results are tabulated by all observed combinations of the columns specified.

Arguments may be used in conjunction with one another.

denominator

Specify this optional argument to change the denominator, e.g. the "N" statistic. Default is NULL. See below for details.

Value

a data frame

Denominators

By default, the ard_categorical() function returns the statistics "n" and "N", where little "n" are the counts for the variable levels, and "N" is the number of non-missing observations. The default calculation for the percentage is merely p = n/N.

However, it is sometimes necessary to provide a different "N" to use as the denominator in this calculation. For example, in a calculation of the rates of various observed adverse events, you may need to update the denominator to the number of enrolled subjects.

In such cases, use the denominator argument to specify a new definition of "N", and subsequently "p". The argument expects a data frame, and the data frame must include the columns specified in ard_categorical(by=) (strata columns are not considered). The updated N and length elements will be updated to be calculated as the number of rows in each combination of the by variables.

Take an example where we need to update the denominator to be subjects enrolled in a trial, e.g. tabulating the number of AEs that occurred within an SOC where some subjects may not have experienced an AE and would not be represented in the ADAE data set. All patients appear in ADSL, however.

ard_categorical(
  data =
    ADAE |>
      dplyr::left_join(ADSL[c("USUBJID", "ARM")], by = "USUBJID") |>
      dplyr::filter(AOCCSFL %in% "Y"),
  by = "ARM",
  variables = "AESOC",
  denominator = ADSL
) |>
  flatten_ard()
#> # A tibble: 141 x 9
#>    group1 group1_level variable variable_level    stat_name stat_label statistic
#>    <chr>  <chr>        <chr>    <chr>             <chr>     <chr>      <chr>    
#>  1 ARM    Placebo      AESOC    CARDIAC DISORDERS n         n          12       
#>  2 ARM    Placebo      AESOC    CARDIAC DISORDERS p         %          0.139534~
#>  3 ARM    Placebo      AESOC    CONGENITAL, FAMI~ n         n          0        
#>  4 ARM    Placebo      AESOC    CONGENITAL, FAMI~ p         %          0        
#>  5 ARM    Placebo      AESOC    EAR AND LABYRINT~ n         n          1        
#>  6 ARM    Placebo      AESOC    EAR AND LABYRINT~ p         %          0.011627~
#>  7 ARM    Placebo      AESOC    EYE DISORDERS     n         n          2        
#>  8 ARM    Placebo      AESOC    EYE DISORDERS     p         %          0.023255~
#>  9 ARM    Placebo      AESOC    GASTROINTESTINAL~ n         n          17       
#> 10 ARM    Placebo      AESOC    GASTROINTESTINAL~ p         %          0.197674~
#> # i 131 more rows
#> # i 2 more variables: warning <chr>, error <chr>

Examples

ard_categorical(ADSL, by = "ARM", variables = "AGEGR1") |>
  flatten_ard()
#> # A tibble: 36 × 9
#>    group1 group1_level variable variable_level stat_name stat_label statistic   
#>    <chr>  <chr>        <chr>    <chr>          <chr>     <chr>      <chr>       
#>  1 ARM    Placebo      AGEGR1   65-80          n         n          42          
#>  2 ARM    Placebo      AGEGR1   65-80          p         %          0.488372093…
#>  3 ARM    Placebo      AGEGR1   <65            n         n          14          
#>  4 ARM    Placebo      AGEGR1   <65            p         %          0.162790697…
#>  5 ARM    Placebo      AGEGR1   >80            n         n          30          
#>  6 ARM    Placebo      AGEGR1   >80            p         %          0.348837209…
#>  7 ARM    Placebo      AGEGR1   NA             N         N          86          
#>  8 ARM    Placebo      AGEGR1   NA             N_obs     N_obs      86          
#>  9 ARM    Placebo      AGEGR1   NA             N_miss    N_miss     0           
#> 10 ARM    Placebo      AGEGR1   NA             N_nonmiss N_nonmiss  86          
#> # ℹ 26 more rows
#> # ℹ 2 more variables: warning <chr>, error <chr>