Simple ARD Statistics

Compute Analysis Results Data (ARD) for simple summary statistics from continuous and categorical data.

Usage

ard_continuous(data, variables, by = NULL, statistics = NULL)

ard_categorical(data, variables, by = NULL, denominator = NULL)

Arguments

data

a data frame

variables

columns to include in summaries. Default is everything().

by

columns to compute statistics by. Default are the columns returned by dplyr::group_vars(data).

statistics

a named list of functions that return a summary statistic, e.g. list(mpg = list(mean = \(x) mean(x, na.rm = TRUE)))

denominator

Specify this argument to change the denominator, e.g. the "N" statistic. Default is NULL. See below for details.

Value

a data frame

Denominators

By default, the ard_categorical() function returns the statistics "n" and "N", where little "n" are the counts for the variable levels, and "N" is the number of non-missing observations. The default calculation for the percentage is merely p = n/N.

However, it is sometimes necessary to provide a different "N" to use as the denominator in this calculation. For example, in a calculation of the rates of various observed adverse events, you may need to update the denominator to the number of subjects who received treatment.

In such cases, use the denominator argument to specify a new definition of "N", and subsequently "p". The argument expects a named list. The names are argument names that will be passed to ard_continuous() to make the variable-wide calculations. When an argument is not specified, then the value passed to ard_categorical() will be utilized.

Take an example where we need to update the denominator to be subjects enrolled in a trial. The argument may look something like this:

ard_categorical(
  data = ADAE,
  by = TRT01P,
  variable = AE
  denominator =
    list(
      data = ADSL,
      statistic = list(N = function(x) sum(!is.na(x)))
    )
)

Note in the above example, no by argument was added to the named list, and therefore, the by argument from the primary call to ard_categorical() will be used. TODO: Update this with a real example for an AE table.

Examples

ard_continuous(mtcars, by = "cyl", variables = c("mpg", "hp"))
#> # A tibble: 36 × 10
#>    group1 group1_level variable stat_name stat_label statistic_fmt_fn statistic
#>  * <chr>  <list>       <chr>    <chr>     <chr>      <list>           <list>   
#>  1 cyl    <dbl [1]>    mpg      N         N          <fn>             <int [1]>
#>  2 cyl    <dbl [1]>    mpg      length    length     <fn>             <int [1]>
#>  3 cyl    <dbl [1]>    mpg      mean      Mean       <fn>             <dbl [1]>
#>  4 cyl    <dbl [1]>    mpg      sd        SD         <fn>             <dbl [1]>
#>  5 cyl    <dbl [1]>    mpg      min       Min        <fn>             <dbl [1]>
#>  6 cyl    <dbl [1]>    mpg      max       Max        <fn>             <dbl [1]>
#>  7 cyl    <dbl [1]>    hp       N         N          <fn>             <int [1]>
#>  8 cyl    <dbl [1]>    hp       length    length     <fn>             <int [1]>
#>  9 cyl    <dbl [1]>    hp       mean      Mean       <fn>             <dbl [1]>
#> 10 cyl    <dbl [1]>    hp       sd        SD         <fn>             <dbl [1]>
#> # ℹ 26 more rows
#> # ℹ 3 more variables: warning <list>, error <list>, context <list>
ard_categorical(mtcars, by = "cyl", variables = c("am", "gear"))
#> # A tibble: 42 × 11
#>    group1 group1_level variable stat_label statistic_fmt_fn variable_level
#>  * <chr>  <list>       <chr>    <chr>      <list>           <list>        
#>  1 cyl    <dbl [1]>    am       n          <fn>             <dbl [1]>     
#>  2 cyl    <dbl [1]>    am       %          <fn>             <dbl [1]>     
#>  3 cyl    <dbl [1]>    am       n          <fn>             <dbl [1]>     
#>  4 cyl    <dbl [1]>    am       %          <fn>             <dbl [1]>     
#>  5 cyl    <dbl [1]>    am       n          <fn>             <dbl [1]>     
#>  6 cyl    <dbl [1]>    am       %          <fn>             <dbl [1]>     
#>  7 cyl    <dbl [1]>    am       n          <fn>             <dbl [1]>     
#>  8 cyl    <dbl [1]>    am       %          <fn>             <dbl [1]>     
#>  9 cyl    <dbl [1]>    am       n          <fn>             <dbl [1]>     
#> 10 cyl    <dbl [1]>    am       %          <fn>             <dbl [1]>     
#> # ℹ 32 more rows
#> # ℹ 5 more variables: warning <list>, error <list>, context <list>,
#> #   stat_name <chr>, statistic <list>