Check Class Data Frame

Check Class Data Frame

Usage

check_class_data_frame(
  x,
  allow_null = FALSE,
  arg_name = rlang::caller_arg(x),
  call = parent.frame()
)

Arguments

x

(object)
object to check

allow_null

(logical(1))
Logical indicating whether a NULL value will pass the test. Default is FALSE

arg_name

(string)
string indicating the label/symbol of the object being checked. Default is rlang::caller_arg(x)

call

The execution environment of a currently running function, e.g. call = caller_env(). The corresponding function call is retrieved and mentioned in error messages as the source of the error.

You only need to supply call when throwing a condition from a helper function which wouldn't be relevant to mention in the message.

Can also be NULL or a defused function call to respectively not display any call or hard-code a code to display.

For more information about error calls, see Including function calls in error messages.