Check Package Installation

check_pkg_installed(): checks whether a package is installed and returns an error or FALSE if not available. If a package search is provided, the function will check whether a minimum version of a package is required.

get_pkg_dependencies() returns a tibble with all dependencies of a specific package.

get_min_version_required() will return, if any, the minimum version of pkg required by reference_pkg, NULL if no minimum version required.

Usage

check_pkg_installed(pkg, reference_pkg = "cards", call = parent.frame())

is_pkg_installed(pkg, reference_pkg = "cards", call = parent.frame())

get_pkg_dependencies(reference_pkg = "cards")

get_all_pkg_dependencies(
  reference_pkg = "cards",
  remove_duplicates = FALSE,
  lib.loc = NULL
)

get_min_version_required(pkg, reference_pkg = "cards")

Arguments

pkg

(string)
Package required

reference_pkg

(string)
the package the function will search for a minimum required version from.

call

(environment)
The execution environment of a currently running function. Default is parent.frame(). This is used to message user about the original function call the resulted in the prompt to install new packages.

remove_duplicates

(logical(1))
if several versions of a package are installed, should only the first one be returned?

lib.loc

(path) location of R library trees to search through, see utils::installed.packages().

Value

logical or error for check_pkg_installed(), NULL or character with the minimum version required for get_min_version_required(), a tibble for get_pkg_dependencies().

Details

get_all_pkg_dependencies() may be used to get the list of dependencies of all installed packages.

Examples

is_pkg_installed("dplyr")
#> [1] TRUE
get_pkg_dependencies()
#> # A tibble: 10 × 6
#>    reference_pkg reference_pkg_version dependency_type pkg       version compare
#>    <chr>         <chr>                 <chr>           <chr>     <chr>   <chr>  
#>  1 cards         0.0.0.9024            Imports         cli       3.6.1   >=     
#>  2 cards         0.0.0.9024            Imports         dplyr     1.1.2   >=     
#>  3 cards         0.0.0.9024            Imports         glue      1.6.2   >=     
#>  4 cards         0.0.0.9024            Imports         rlang     1.1.1   >=     
#>  5 cards         0.0.0.9024            Imports         tidyr     1.3.0   >=     
#>  6 cards         0.0.0.9024            Imports         tidysele… 1.2.0   >=     
#>  7 cards         0.0.0.9024            Depends         R         4.1     >=     
#>  8 cards         0.0.0.9024            Suggests        spelling  2.2.0   >=     
#>  9 cards         0.0.0.9024            Suggests        testthat  3.2.0   >=     
#> 10 cards         0.0.0.9024            Suggests        withr     3.0.0   >=     
get_min_version_required("brms")
#> NULL