ARD Attributes

Add variable attributes to an ARD data frame.

The label attribute will be added for all columns, and when no label is specified and no label has been set for a column using the label= argument, the column name will be placed in the label statistic.

The class attribute will also be returned for all columns.

Any other attribute returned by attributes() will also be added, e.g. factor levels.

Usage

ard_attributes(data, variables = everything(), label = NULL)

Arguments

data

a data frame

variables

variable to include

label

named list of variable labels, e.g. list(mpg = "MPG"). Default is NULL

Value

a data frame

Examples

df <- dplyr::tibble(var1 = letters, var2 = LETTERS)
attr(df$var1, 'label') <- "Lowercase Letters"

ard_attributes(df, variables = everything())
#> # A tibble: 4 × 5
#>   variable stat_name statistic stat_label     context   
#> * <chr>    <chr>     <list>    <chr>          <chr>     
#> 1 var1     label     <chr [1]> Variable Label attributes
#> 2 var1     class     <chr [1]> Variable Class attributes
#> 3 var2     label     <chr [1]> Variable Label attributes
#> 4 var2     class     <chr [1]> Variable Class attributes