Continuous ARD Statistics

Compute Analysis Results Data (ARD) for simple continuous summary statistics.

Usage

ard_continuous(
  data,
  variables,
  by = NULL,
  strata = NULL,
  statistics = everything() ~ continuous_variable_summary_fns(),
  fmt_fn = NULL,
  stat_labels = everything() ~ default_stat_labels()
)

Arguments

data

(data.frame)
a data frame

variables

(tidy-select)
columns to include in summaries

by, strata

(tidy-select)
columns to by/stratified by for summary statistic calculation. Arguments are similar, but with an important distinction:

by: results are calculated by all combinations of the columns specified, including unobserved combinations and unobserved factor levels.

strata: results are calculated by all observed combinations of the columns specified.

Arguments may be used in conjunction with one another.

statistics

(formula-list-selector)
a named list, a list of formulas, or a single formula where the list element is a named list of functions (or the RHS of a formula), e.g. list(mpg = list(mean = \(x) mean(x))).

The value assigned to each variable must also be a named list, where the names are used to reference a function and the element is the function object. Typically, this function will return a scalar statistic, but a function that returns a named list of results is also acceptable, e.g. list(conf.low = -1, conf.high = 1). However, when errors occur, the messaging will be less clear in this setting.

fmt_fn

(formula-list-selector)
a named list, a list of formulas, or a single formula where the list element is a named list of functions (or the RHS of a formula), e.g. list(mpg = list(mean = \(x) round(x, digits = 2) |> as.character)).

stat_labels

(formula-list-selector)
a named list, a list of formulas, or a single formula where the list element is either a named list or a list of formulas defining the statistic labels, e.g. everything() ~ list(mean = "Mean", sd = "SD") or everything() ~ list(mean ~ "Mean", sd ~ "SD").

Value

a data frame

Examples

ard_continuous(ADSL, by = "ARM", variables = "AGE") |>
  flatten_ard()
#> # A tibble: 24 × 8
#>    group1 group1_level     variable stat_name stat_label statistic warning error
#>    <chr>  <chr>            <chr>    <chr>     <chr>      <chr>     <chr>   <chr>
#>  1 ARM    Placebo          AGE      N         N          86        NA      NA   
#>  2 ARM    Placebo          AGE      mean      Mean       75.20930… NA      NA   
#>  3 ARM    Placebo          AGE      sd        SD         8.590167… NA      NA   
#>  4 ARM    Placebo          AGE      median    Median     76        NA      NA   
#>  5 ARM    Placebo          AGE      p25       25th Perc… 69        NA      NA   
#>  6 ARM    Placebo          AGE      p75       75th Perc… 82        NA      NA   
#>  7 ARM    Placebo          AGE      min       Min        52        NA      NA   
#>  8 ARM    Placebo          AGE      max       Max        89        NA      NA   
#>  9 ARM    Xanomeline High… AGE      N         N          84        NA      NA   
#> 10 ARM    Xanomeline High… AGE      mean      Mean       74.38095… NA      NA   
#> # ℹ 14 more rows