Hierarchical ARD Statistics

Performs hierarchical or nested tabulations, e.g. tabulates AE terms nested within AE system organ class.

ard_hierarchical() includes summaries for the last variable listed in the variables argument, nested within the other variables included.

ard_hierarchical_count() includes summaries for all variables listed in the variables argument each summary nested within the preceding variables, e.g. variables=c(AESOC, AETERM) summarizes AETERM nested in AESOC, and also summarizes the counts of AESOC.

Usage

ard_hierarchical(
  data,
  variables,
  by = NULL,
  statistics = everything() ~ categorical_variable_summary_fns(),
  denominator = NULL,
  fmt_fn = NULL,
  stat_labels = everything() ~ default_stat_labels()
)

ard_hierarchical_count(
  data,
  variables,
  by = NULL,
  fmt_fn = NULL,
  stat_labels = everything() ~ default_stat_labels()
)

Arguments

data

(data.frame)
a data frame

variables

(tidy-select)
variables to perform the nested/hierarchical tabulations within.

by

(tidy-select)
variables to perform tabulations by. All combinations of the variables specified here appear in results. Default is NULL

statistics

(formula-list-selector)
a named list, a list of formulas, or a single formula where the list element is a named list of functions (or the RHS of a formula), e.g. list(mpg = categorical_variable_summary_fns()).

denominator

(data.frame)
Specify this optional argument to change the denominator, e.g. the "N" statistic. Default is NULL. See below for details.

fmt_fn

(formula-list-selector)
a named list, a list of formulas, or a single formula where the list element is a named list of functions (or the RHS of a formula), e.g. list(mpg = list(mean = \(x) round(x, digits = 2) |> as.character)).

stat_labels

(formula-list-selector)
a named list, a list of formulas, or a single formula where the list element is either a named list or a list of formulas defining the statistic labels, e.g. everything() ~ list(n = "n", p = "pct") or everything() ~ list(n ~ "n", p ~ "pct").

Value

an ARD data frame

Examples

ard_hierarchical(
  data = ADAE,
  variables = c(AESOC, AETERM),
  by = c(TRTA, AESEV),
  denominator = ADSL |> dplyr::rename(TRTA = ARM)
) |>
  flatten_ard()
#> # A tibble: 4,365 × 13
#>    group1 group1_level group2 group2_level group3 group3_level      variable
#>    <chr>  <chr>        <chr>  <chr>        <chr>  <chr>             <chr>   
#>  1 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#>  2 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#>  3 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#>  4 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#>  5 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#>  6 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#>  7 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#>  8 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#>  9 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#> 10 TRTA   Placebo      AESEV  MILD         AESOC  CARDIAC DISORDERS AETERM  
#> # ℹ 4,355 more rows
#> # ℹ 6 more variables: variable_level <chr>, stat_name <chr>, stat_label <chr>,
#> #   statistic <chr>, warning <chr>, error <chr>

ard_hierarchical_count(
  data = ADAE,
  variables = c(AESOC, AETERM),
  by = TRTA
) |>
  flatten_ard()
#> # A tibble: 795 × 11
#>    group1 group1_level group2 group2_level variable variable_level     stat_name
#>    <chr>  <chr>        <chr>  <chr>        <chr>    <chr>              <chr>    
#>  1 TRTA   Placebo      NA     NA           AESOC    CARDIAC DISORDERS  n        
#>  2 TRTA   Placebo      NA     NA           AESOC    CONGENITAL, FAMIL… n        
#>  3 TRTA   Placebo      NA     NA           AESOC    EAR AND LABYRINTH… n        
#>  4 TRTA   Placebo      NA     NA           AESOC    EYE DISORDERS      n        
#>  5 TRTA   Placebo      NA     NA           AESOC    GASTROINTESTINAL … n        
#>  6 TRTA   Placebo      NA     NA           AESOC    GENERAL DISORDERS… n        
#>  7 TRTA   Placebo      NA     NA           AESOC    HEPATOBILIARY DIS… n        
#>  8 TRTA   Placebo      NA     NA           AESOC    IMMUNE SYSTEM DIS… n        
#>  9 TRTA   Placebo      NA     NA           AESOC    INFECTIONS AND IN… n        
#> 10 TRTA   Placebo      NA     NA           AESOC    INJURY, POISONING… n        
#> # ℹ 785 more rows
#> # ℹ 4 more variables: stat_label <chr>, statistic <chr>, warning <chr>,
#> #   error <chr>