Process tidyselectors

Functions processes tidyselect arguments passed to functions in the cards package. The processed values saved to the calling environment, by default.

process_selectors(): the arguments will be processed with tidyselect and converted to a vector of character column names.

process_formula_selectors(): for arguments that expect named lists or lists of formulas (where the LHS of the formula is a tidyselector). This function processes these inputs and returns a named list. If an name is repeated, the last entry is kept.

fill_formula_selectors(): when users override the default argument values, it can be important to ensure that each column from a data frame is assigned a value. This function checks that each column in data has an assigned value, and if not, fills the value in with the default value passed here.

compute_formula_selector(): used in process_formula_selectors() to evaluate a single argument.

check_list_elements(): accepts named arguments where the name is a list that exists in the env, and the argument value is a predicate function used to the values of the list.

Usage

process_selectors(data, ..., env = rlang::caller_env())

process_formula_selectors(data, ..., env = rlang::caller_env())

fill_formula_selectors(data, ..., env = rlang::caller_env())

compute_formula_selector(
  data,
  x,
  arg_name = rlang::caller_arg(x),
  env = rlang::caller_env()
)

check_list_elements(..., error_msg = list(), env = rlang::caller_env())

Arguments

data

a data frame

...

named arguments where the value of the argument is processed with tidyselect.

process_selectors(): the values are tidyselect-compatible selectors

process_formula_selectors(): the values are named lists, list of formulas a combination of both, or a single formula. Users may pass ~value as a shortcut for everything() ~ value.

check_list_elements(): named arguments where the name matches an existing list in the env environment, and the value is a predicate function to test each element of the list, e.g. each element must be a string or a function.

env

env to save the results to. Default is the calling environment.

x

a named list, list of formulas, or a single formula that will be converted to a named list.

arg_name

a string with the argument named being processed. Used in error messaging.

error_msg

a named list where the list elements are strings that will be used in error messaging when mis-specified arguments are passed. Elements "{arg_name}" and "{variable}" are available using glue syntax for messaging.

Examples

example_env <- rlang::new_environment()

process_selectors(ADSL, variables = starts_with("TRT"), env = example_env)
get(x = "variables", envir = example_env)
#> [1] "TRT01P"  "TRT01PN" "TRT01A"  "TRT01AN" "TRTSDT"  "TRTEDT"  "TRTDUR" 

process_formula_selectors(
  ADSL,
  statistics = list(starts_with("TRT") ~ mean, TRTSDT = min),
  env = example_env
)
get(x = "statistics", envir = example_env)
#> $TRT01P
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x561088f5b2b0>
#> <environment: namespace:base>
#> 
#> $TRT01PN
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x561088f5b2b0>
#> <environment: namespace:base>
#> 
#> $TRT01A
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x561088f5b2b0>
#> <environment: namespace:base>
#> 
#> $TRTSDT
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x561088f5b2b0>
#> <environment: namespace:base>
#> 
#> $TRTEDT
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x561088f5b2b0>
#> <environment: namespace:base>
#> 
#> $TRTDUR
#> function (x, ...) 
#> UseMethod("mean")
#> <bytecode: 0x561088f5b2b0>
#> <environment: namespace:base>
#> 
#> $TRTSDT
#> function (..., na.rm = FALSE)  .Primitive("min")
#> 

# process one list
compute_formula_selector(ADSL, x = starts_with("U") ~ 1L)
#> $USUBJID
#> [1] 1
#>