Summary Functions

continuous_variable_summary_fns() returns a named list of summary functions for continuous variables. Some functions include slight modifications to their base equivalents. For example, the min() and max() functions return NA instead of Inf when an empty vector is passed.

categorical_variable_summary_fns() summary functions for categorical variables.

missing_variable_summary_fns() summary functions suitable for variable-level summaries, such as number and rate of missing data.

Usage

categorical_variable_summary_fns(summaries = c("n", "p", "N"))

continuous_variable_summary_fns(
  summaries = c("N", "mean", "sd", "median", "p25", "p75", "min", "max")
)

missing_variable_summary_fns(
  summaries = c("N_obs", "N_miss", "N_nonmiss", "p_miss", "p_nonmiss")
)

Arguments

summaries

(character)
a character vector of results to include in output.

categorical_variable_summary_fns(): Select one or more from 'N', 'n', 'p'

continuous_variable_summary_fns(): Select one or more from 'N', 'mean', 'sd', 'median', 'p25', 'p75', 'min', 'max'.

missing_variable_summary_fns(): Select one or more from 'N_obs', 'N_miss', 'N_nonmiss', 'p_miss', 'p_nonmiss'.

Value

named list of summary functions

Examples

# continuous variable summaries
ard_continuous(
  ADSL,
  variables = "AGE",
  statistics = ~continuous_variable_summary_fns(c("N", "median"))
) |>
  flatten_ard()
#> # A tibble: 2 × 6
#>   variable stat_name stat_label statistic warning error
#>   <chr>    <chr>     <chr>      <chr>     <chr>   <chr>
#> 1 AGE      N         N          254       NA      NA   
#> 2 AGE      median    Median     77        NA      NA   

# categorical variable summaries
ard_categorical(
  ADSL,
  variables = "AGEGR1",
  statistics = ~categorical_variable_summary_fns(c("n", "N"))
) |>
  flatten_ard()
#> # A tibble: 4 × 7
#>   variable variable_level stat_name stat_label statistic warning error
#>   <chr>    <chr>          <chr>     <chr>      <chr>     <chr>   <chr>
#> 1 AGEGR1   65-80          n         n          144       NA      NA   
#> 2 AGEGR1   <65            n         n          33        NA      NA   
#> 3 AGEGR1   >80            n         n          77        NA      NA   
#> 4 AGEGR1   NA             N         N          254       NA      NA   

# summary for rates of missing data
ard_missing(
  ADSL,
  variables = c("AGE", "AGEGR1"),
  statistics = ~missing_variable_summary_fns()
) |>
  flatten_ard()
#> # A tibble: 10 × 6
#>    variable stat_name stat_label    statistic warning error
#>    <chr>    <chr>     <chr>         <chr>     <chr>   <chr>
#>  1 AGE      N_obs     Vector Length 254       NA      NA   
#>  2 AGE      N_miss    N Missing     0         NA      NA   
#>  3 AGE      N_nonmiss N Non-missing 254       NA      NA   
#>  4 AGE      p_miss    % Missing     0         NA      NA   
#>  5 AGE      p_nonmiss % Non-missing 1         NA      NA   
#>  6 AGEGR1   N_obs     Vector Length 254       NA      NA   
#>  7 AGEGR1   N_miss    N Missing     0         NA      NA   
#>  8 AGEGR1   N_nonmiss N Non-missing 254       NA      NA   
#>  9 AGEGR1   p_miss    % Missing     0         NA      NA   
#> 10 AGEGR1   p_nonmiss % Non-missing 1         NA      NA